la derniere modification et l'ajout des autre fonctionnes

+++ backend/database/migrations/2026_02_02_223235_create_employes_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('employes', function (Blueprint $table) {
-            $table->integer('Doti')->primary();
+            $table->unsignedBigInteger('Doti')->primary(); 
             $table->string('nom');
             $table->string('prenom');
             $table->string('CIN')->unique();
